[Fix] Refactor core package
