+++ app/.gitignore
@@ -1,1 +1,3 @@
 /build
+src/main/res/values/keys.xml
+google-services.json
