+++ core/build.gradle
@@ -36,6 +36,8 @@ android {
 
 dependencies {
     with defaultDependencies
+    implementation 'androidx.security:security-crypto-ktx:1.1.0-alpha06'
+    api 'com.google.code.gson:gson:2.10.1'
 }
 
 with jitpackPublisher
