+++ credentialsauth/build.gradle
@@ -38,14 +38,12 @@ dependencies {
     implementation project(':core')
     with defaultDependencies
 
-    implementation 'com.google.android.material:material:1.8.0'
-    implementation 'androidx.security:security-crypto-ktx:1.1.0-alpha05'
+    implementation 'com.google.android.material:material:1.9.0'
     api 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation 'com.google.code.gson:gson:2.10'
     implementation 'com.squareup.okhttp3:logging-interceptor:4.10.0'
     testImplementation 'com.squareup.okhttp3:mockwebserver:4.10.0'
-    testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4'
+    testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1'
 }
 
 with jitpackPublisher
