+++ credentialsauth/src/main/java/com/sun/auth/credentials/repositories/local/AuthLocalDataSource.kt
@@ -1,6 +1,6 @@
 package com.sun.auth.credentials.repositories.local
 
-import com.sun.auth.credentials.repositories.local.api.SharedPrefApi
+import com.sun.auth.core.SharedPrefApi
 import com.sun.auth.credentials.repositories.model.AuthToken
 import com.sun.auth.credentials.utils.PREF_AUTH_TOKEN
