+++ googleauth/src/main/java/com/sun/auth/google/standard/AuthClient.kt
@@ -105,7 +105,7 @@ internal class AuthClient(
                 signInCallback?.onResult(data = account)
             }
         } catch (e: ApiException) {
-            if (e.statusCode == Activity.RESULT_CANCELED) {
+            if (e.status.isCanceled) {
                 signInCallback?.onResult(error = CancellationAuthException())
             } else {
                 signInCallback?.onResult(error = SocialAuthException(e))
