+++ googlefirebaseauth/build.gradle
@@ -40,7 +40,7 @@ dependencies {
 
     api platform('com.google.firebase:firebase-bom:31.1.1')
     api 'com.google.firebase:firebase-auth-ktx'
-    api 'com.google.android.gms:play-services-auth:20.4.1'
+    api 'com.google.android.gms:play-services-auth:20.5.0'
 }
 
 with jitpackPublisher
